+++ app/globals.css
@@ -9,7 +9,7 @@
         --foreground: 224 71.4% 4.1%;
         --card: 0 0% 100%;
         --card-foreground: 224 71.4% 4.1%;
-        --popover: 0 0% 100%;
+        --popover: 0 0% 100%; 
         --popover-foreground: 224 71.4% 4.1%;
         --primary: 262.1 83.3% 57.8%;
         --primary-foreground: 210 20% 98%;
